+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog 1.0.0].
 ## [Unreleased]
 
 - New versions of a document created with `insert_document_xml` can now be annotated
+- BREAKING: Renamed `save_judgment_xml` to `update_document_xml`
 
 ## [Release 15.1.2
 - Expose the creation date of a version